implement create classification endpoint and change top p param type in completion form.

+++ src/main/java/com/openai/domain/CompletionForm.java
@@ -14,7 +14,7 @@ public class CompletionForm {
     private double temperature;
 
     @JsonProperty("top_p")
-    private int topP;
+    private double topP;
 
     @JsonProperty("n")
     private int n;
